Committed a module containing functions for implementing a sliding-window-look-forward-average smoothing algorithm.
This was created because existing trimming programs didn't do _exactly_
what we needed.

I wouldn't really advise that people use this much. It operates, in my
context, as a part of Consed.pm and works well but it is a bit...
special. :)

The rest of the files comprise a consed project and are used to test
Trim.pm. I really should rewrite t/consed.t to use this as well but
t/data/consed_project represents a minimal project and
t/acefile.ace.1, along with t/acefile.singlets is a much larger project.

I did not include the chromats (too big) _BUT_ the phreds in
t/data/consed_project/phd_dir could be used as fodder for
SeqIO/csmscf.pm to create synthetic chromats to put in that directory.

In fact, that is _exactly_ why I wrote csmscf.pm. It really is good for
_something_!

svn path=/bioperl-live/trunk/; revision=2855